Started code refactorization; Added new folders for models, agents, logs, and weights
