--- src/model_A2C_PPO/.gitignore
@@ -1,7 +1,0 @@
-/logs
-/model_checkpoints
-*.py~
-*.py.un~
-*.py.swp
-*.un~
-/weights
